# api/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_noah-isme_sma-adp-api_internal_models.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_service.CreateAnnouncementRequest:
    properties:
      audience:
        type: string
      content:
        type: string
      created_by:
        type: string
      expires_at:
        type: string
      is_pinned:
        type: boolean
      priority:
        type: string
      published_at:
        type: string
      target_class_id:
        type: string
      title:
        type: string
    required:
    - audience
    - content
    - created_by
    - priority
    - published_at
    - title
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_service.CreateClassRequest:
    properties:
      grade:
        type: string
      homeroom_teacher_id:
        type: string
      name:
        type: string
      track:
        type: string
    required:
    - grade
    - name
    - track
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_service.UpdateAnnouncementRequest:
    properties:
      audience:
        type: string
      content:
        type: string
      expires_at:
        type: string
      is_pinned:
        type: boolean
      priority:
        type: string
      published_at:
        type: string
      target_class_id:
        type: string
      title:
        type: string
    required:
    - audience
    - content
    - priority
    - published_at
    - title
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_service.UpdateClassRequest:
    properties:
      grade:
        type: string
      homeroom_teacher_id:
        type: string
      name:
        type: string
      track:
        type: string
    required:
    - grade
    - name
    - track
    type: object
...
  github_com_noah-isme_sma-adp-api_pkg_errors.Error:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  github_com_noah-isme_sma-adp-api_pkg_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_errors.Error'
      meta:
        additionalProperties: true
        type: object
      pagination:
        $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_models.Pagination'
    type: object
info:
  contact: {}
  description: Bootstrap server for Golang migration (Phase 0)
  title: SMA ADP API
  version: 0.1.0
paths:
  /analytics/attendance:
    get:
      parameters:
      - description: Term ID
        in: query
        name: term_id
        required: true
        type: string
      - description: Class ID
        in: query
        name: class_id
        required: true
        type: string
      - description: Date From (RFC3339)
        in: query
        name: date_from
        type: string
      - description: Date To (RFC3339)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Attendance analytics
      tags:
      - Analytics
  /analytics/behavior:
    get:
      parameters:
      - description: Term ID
        in: query
        name: term_id
        required: true
        type: string
      - description: Student ID
        in: query
        name: student_id
        required: true
        type: string
      - description: Class ID
        in: query
        name: class_id
        required: true
        type: string
      - description: Date From (RFC3339)
        in: query
        name: date_from
        type: string
      - description: Date To (RFC3339)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Behavior analytics
      tags:
      - Analytics
  /analytics/grades:
    get:
      parameters:
      - description: Term ID
        in: query
        name: term_id
        required: true
        type: string
      - description: Class ID
        in: query
        name: class_id
        required: true
        type: string
      - description: Subject ID
        in: query
        name: subject_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Grade analytics
      tags:
      - Analytics
  /analytics/system:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: System analytics
      tags:
      - Analytics
  /announcements:
    get:
      parameters:
      - description: Comma-separated audience roles
        in: query
        name: audience
        type: string
      - description: Comma-separated class IDs
        in: query
        name: classIds
        type: string
      - description: Include pinned announcements
        in: query
        name: includePinned
        type: boolean
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List announcements
      tags:
      - Announcements
    post:
      consumes:
      - application/json
      parameters:
      - description: Announcement payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_service.CreateAnnouncementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create announcement
      tags:
      - Announcements
  /announcements/{id}:
    delete:
      parameters:
      - description: Announcement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete announcement
      tags:
      - Announcements
    get:
      parameters:
      - description: Announcement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get announcement
      tags:
      - Announcements
    put:
      consumes:
      - application/json
      parameters:
      - description: Announcement ID
        in: path
        name: id
        required: true
        type: string
      - description: Announcement payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_service.UpdateAnnouncementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update announcement
      tags:
      - Announcements
...
  /classes:
    get:
      parameters:
      - description: Filter by grade
        in: query
        name: grade
        type: string
      - description: Filter by track
        in: query
        name: track
        type: string
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List classes
      tags:
      - Classes
    post:
      consumes:
      - application/json
      parameters:
      - description: Class payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_service.CreateClassRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create class
      tags:
      - Classes
  /classes/{id}:
    delete:
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete class
      tags:
      - Classes
    get:
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get class detail
      tags:
      - Classes
    put:
      consumes:
      - application/json
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      - description: Class payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_service.UpdateClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update class
      tags:
      - Classes
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"